updated components for better overview
