+++ frontend/src/App.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import './App.css';
 import LandingPage from "./pages/LandingPage";
 import {BrowserRouter, Route, Routes} from "react-router-dom";
-import PictureOfTheDay from "./components/PictureOfTheDay";
-import RandomPicture from "./components/RandomPicture";
+import PictureOfTheDayPage from "./pages/PictureOfTheDayPage";
+import RandomPicturePage from "./pages/RandomPicturePage";
 import ArchivePage from "./pages/ArchivePage";
 import FavouritesPage from "./pages/FavouritesPage";
 
@@ -16,9 +16,9 @@ function App() {
         <Route path="/"
                element={<LandingPage/>}/>
         <Route path="/picoftheday"
-               element={<PictureOfTheDay/>}/>
+               element={<PictureOfTheDayPage/>}/>
         <Route path="/randompicture"
-               element={<RandomPicture/>}/>
+               element={<RandomPicturePage/>}/>
         <Route path="/archive"
                element={<ArchivePage/>}/>
         <Route path="/favourites"
