+++ frontend/src/hooks/useArchive.ts
@@ -4,7 +4,6 @@ import {getArchive} from "../service/api-service";
 
 export default function useArchive(){
     const [archivePictures, setArchivePictures] = useState<NasaPicture[]>();
-
     useEffect(() =>{
         getArchive()
             .then(archivePictures => setArchivePictures(archivePictures))
